requirements.txt
python-binance==1.0.19
openai>=1.0.0
numpy>=1.24.0
requests>=2.28.0